Fix CI change detection for master commits
It seems that, until now, we've been comparing HEAD to itself for master commits. This differentiates the diff logic between pull requests and master commits (+6 squashed commits)
Squashed commits:
[8c0cbc90] Refactor installation to use stashing strategy

Refactor installation to use stashing strategy:

- when installing A, install it into A's stash, then move A's node_modules up to top level
- when installing B, stash A's node_modules into A's stash, install into B's stash, then move B's node_modules up to top level
- from then on, unstash whichever one is requested by a `jazelle install` call

This strategy optimizes for fast context-switching between projects, assuming no yarn.lock changes between switches

Previously the installation strategy attempted to reuse the existing node_modules regardless of what was in it, but that approach turned out to create cases where the wrong version of things would be installed after context-switching between projects with different (but compatible) dep versions. This change isolates node_modules per project so that there's never cross-contamination

This should avoid the need of running `jazelle prune` when context-switching
[25485d96] Fix check command when no version policy

Previously, if no version policy was defined, `jazelle check` would erroneously report no discrepancies (the determination logic was correct, but the logging logic was not)

As a fly-by, this also adds a `--json` argument so that `jazelle check` can be piped to jq (e.g. `jazelle check --json | jq .jest` to filter the report to only `jest` version discrepancies
[6120f7d7] Add jest configuration option

This change is enabling test configuration for typescript projects.
[c070f223] Add basic configuration to enable typescript support
[d9b3e7b7] Update babel worker to load fusion babel config

This fixes an issue where functions in the babel config were lost
in the serialization of arguments to the worker. This also adds a
regression test.
[37ac2a9e] Add / fix various things

- Add reporting of effective vs system jazelle version in `jazelle version` command (to help debug)
- Add `jazelle node` command
- Fix cwd handling of `jazelle yarn` command
- Fix PATH order in `jazelle dev` (fixes machine-specific fusion-cli compilation error cc @angus-c)